+++ pyproject.toml
@@ -4,7 +4,7 @@ description = "Make websites accessible for AI agents"
 authors = [
     { name = "Gregor Zunic" }
 ]
-version = "0.1.22"
+version = "0.1.24"
 readme = "README.md"
 requires-python = ">=3.11"
 classifiers = [
@@ -17,7 +17,7 @@ dependencies = [
     "beautifulsoup4>=4.12.3",
     "httpx>=0.27.2",
     "langchain>=0.3.14",
-    "langchain-openai>=0.2.14",
+    "langchain-openai>=0.3.0",
     "langchain-anthropic>=0.3.1",
     "langchain-fireworks>=0.2.5",
     "langchain-aws>=0.2.10",
